refactored storage handling
